# perf(ebpf): improve decodeEvents