### 1. Explain what the PR does

3b8c0f615 **perf(ebpf): improve NUL byte trimming**

```
Introduce utils.TrimTrailingNUL function to handle NUL byte trimming
in a more efficient manner.

Running tool: /home/gg/.goenv/versions/1.24.0/bin/go test -benchmem
-run=^$ -tags ebpf -bench
^(BenchmarkBytesTrimRight_WithNUL|BenchmarkBytesTrimRight_NoNUL|
BenchmarkTrimTrailingNUL_WithNUL|BenchmarkTrimTrailingNUL_NoNUL)$
github.com/aquasecurity/tracee/pkg/utils -benchtime=200000000x -count=5 timeout=5m

goos: linux
goarch: amd64
pkg: github.com/aquasecurity/tracee/pkg/utils
cpu: AMD Ryzen 9 7950X 16-Core Processor

| Benchmark               | Avg Time (ns/op) | Improvement vs TrimRight |
|-------------------------|------------------|--------------------------|
| TrimTrailingNUL_WithNUL | 1.980 ns         | 17.1% faster             |
| BytesTrimRight_WithNUL  | 2.388 ns         | —                        |
| TrimTrailingNUL_NoNUL   | 1.311 ns         | 26.9% faster             |
| BytesTrimRight_NoNUL    | 1.794 ns         | —                        |
```

d7861e6c5 **perf(ebpf): improve decodeEvents (HashTaskID)**

```
Metrics from local benchmarking on a AMD Ryzen 9 7950X 16-Core Processor
with 64GB RAM running 6.12.25-1-MANJARO. The stress test was run
with 8mi events triggered in 8 threads.

- CPU time

| Function     | Before    | After     | Change    | Improvement (%) |
|--------------|-----------|-----------|-----------|-----------------|
| decodeEvents | 26.87 sec | 25.22 sec | -1.65 sec | 6.14%           |
| HashTaskID   | 0.38 sec  | 0.34 sec  | -0.04 sec | 10.53%          |

- Memory

| Metric        | Function     | Before  | After   | Change   | Improv. (%) |
|---------------|--------------|---------|---------|----------|-------------|
| alloc_space   | decodeEvents | 8.52 GB | 8.48 GB | -0.04 GB | 0.47%       |
| inuse_objects | decodeEvents | 1.17 K  | 0.85 K  | -0.32 K  | 27.35%      |
| alloc_objects | decodeEvents | 91.96 M | 89.31 M | -2.65 M  | 2.88%       |
```

### 2. Explain how to test it

<!--
Maintainer will review the code, and test the fix/feature, how to run Tracee ?
Give a full command line example and what to look for.
-->

### 3. Other comments

<!--
Links? References? Anything pointing to more context about the change.
-->